+++ packages/magento-cart-shipping-address/components/CustomerAddressForm/CustomerAddressForm.tsx
@@ -14,7 +14,7 @@ import { CustomerDocument } from '@graphcommerce/magento-customer'
 import { ActionCardListForm, filterNonNullableKeys } from '@graphcommerce/next-ui'
 import { i18n } from '@lingui/core'
 import { Box, SxProps, Theme } from '@mui/material'
-import React, { useMemo } from 'react'
+import React, { useEffect } from 'react'
 import { findCustomerAddressFromCartAddress } from '../../utils/findCustomerAddressFromCartAddress'
 import { GetAddressesDocument } from '../ShippingAddressForm/GetAddresses.gql'
 import { CustomerAddressActionCard } from './CustomerAddressActionCard'
@@ -105,20 +105,26 @@ export function CustomerAddressForm(props: CustomerAddressListProps) {
   const { handleSubmit, error, control, setValue, watch } = form
   const formAddressId = watch('customer_address_id')
 
-  useMemo(() => {
+  useEffect(() => {
     if (mode === 'both' || mode === 'shipping') {
-      if (!cartAddressId && defaultShippingId) {
-        // console.log('shippingAddress.customer_address_id', defaultShippingId)
+      if (!cartAddressId && !cartShipping && defaultShippingId) {
         setValue('customer_address_id', defaultShippingId, { shouldValidate: true })
       }
     }
     if (mode === 'billing') {
-      if (!cartAddressId && defaultBillingId) {
-        // console.log('billingAddress.customer_address_id', defaultBillingId)
+      if (!cartAddressId && !cartBilling && defaultBillingId) {
         setValue('customer_address_id', defaultBillingId, { shouldValidate: true })
       }
     }
-  }, [cartAddressId, defaultBillingId, defaultShippingId, mode, setValue])
+  }, [
+    cartAddressId,
+    cartBilling,
+    cartShipping,
+    defaultBillingId,
+    defaultShippingId,
+    mode,
+    setValue,
+  ])
 
   const submit = handleSubmit(() => {})
 
@@ -129,7 +135,6 @@ export function CustomerAddressForm(props: CustomerAddressListProps) {
 
   return (
     <>
-      <FormPersist form={form} name='CustomerAddressForm' />
       <FormAutoSubmit control={form.control} submit={submit} wait={0} />
       <Box component='form' onSubmit={submit} noValidate sx={sx}>
         <ActionCardListForm
@@ -148,6 +153,7 @@ export function CustomerAddressForm(props: CustomerAddressListProps) {
         <ApolloCartErrorAlert error={error} />
       </Box>
       {formAddressId === -1 && children}
+      <FormPersist form={form} name='CustomerAddressForm' />
     </>
   )
 }