+++ packages/magento-cart-shipping-address/components/ShippingAddressForm/ShippingAddressForm.tsx
@@ -1,4 +1,5 @@
 import {
+  CheckboxElement,
   FormAutoSubmit,
   FormPersist,
   TextFieldElement,
@@ -18,6 +19,7 @@ import {
   CustomerDocument,
   NameFields,
   useCustomerQuery,
+  CompanyFields,
 } from '@graphcommerce/magento-customer'
 import { CountryRegionsDocument, StoreConfigDocument } from '@graphcommerce/magento-store'
 import { Form, FormRow } from '@graphcommerce/next-ui'
@@ -83,7 +85,7 @@ export const ShippingAddressForm = React.memo<ShippingAddressFormProps>((props) 
 
   const form = useFormGqlMutationCart(Mutation, {
     defaultValues: isCartAddressACustomerAddress(customerQuery?.customer?.addresses, currentAddress)
-      ? { saveInAddressBook: true }
+      ? { saveInAddressBook: true, isCompany: false }
       : {
           // todo(paales): change to something more sustainable
           firstname: currentAddress?.firstname ?? customerQuery?.customer?.firstname ?? '',
@@ -92,6 +94,7 @@ export const ShippingAddressForm = React.memo<ShippingAddressFormProps>((props) 
             currentAddress?.telephone !== '000 - 000 0000' ? currentAddress?.telephone : '',
           city: currentAddress?.city ?? '',
           company: currentAddress?.company ?? '',
+          vatId: currentAddress?.vat_id ?? '',
           postcode: currentAddress?.postcode ?? '',
           street: currentAddress?.street?.[0] ?? '',
           houseNumber: currentAddress?.street?.[1] ?? '',
@@ -99,6 +102,7 @@ export const ShippingAddressForm = React.memo<ShippingAddressFormProps>((props) 
           regionId: currentAddress?.region?.region_id ?? null,
           countryCode: currentAddress?.country.code ?? shopCountry, // todo: replace by the default shipping country of the store + geoip,
           saveInAddressBook: true,
+          isCompany: Boolean(currentAddress?.company || currentAddress?.vat_id),
         },
     mode: 'onChange',
     experimental_useV2: true,
@@ -129,9 +133,11 @@ export const ShippingAddressForm = React.memo<ShippingAddressFormProps>((props) 
         control={form.control}
         name={['postcode', 'countryCode', 'regionId']}
       />
-      <FormPersist form={form} name='ShippingAddressForm' />
+
+      <CompanyFields form={form} />
       <NameFields form={form} />
       <AddressFields form={form} />
+
       <FormRow>
         <TextFieldElement
           control={form.control}
@@ -146,7 +152,16 @@ export const ShippingAddressForm = React.memo<ShippingAddressFormProps>((props) 
           showValid
         />
       </FormRow>
+      {customerQuery?.customer && (
+        <CheckboxElement
+          control={form.control}
+          name='saveInAddressBook'
+          label={<Trans id='Save in address book' />}
+        />
+      )}
+
       <ApolloCartErrorAlert error={error} />
+      <FormPersist form={form} name='ShippingAddressForm' />
     </Form>
   )
 })