+++ packages/next-ui/ActionCard/ActionCard.tsx
@@ -37,9 +37,9 @@ export type ActionCardProps = {
   price?: React.ReactNode
   after?: React.ReactNode
   secondaryAction?: React.ReactNode
-  onClick?: (event: React.MouseEvent<HTMLElement>, value: string | number | null) => void
+  onClick?: (event: React.MouseEvent<HTMLElement>, value: string | number | boolean | null) => void
   selected?: boolean
-  value: string | number | null
+  value: string | number | boolean | null
   reset?: React.ReactNode
   disabled?: boolean
   error?: boolean